+++ src/core/data/modules/base.ts
@@ -9,6 +9,7 @@
 import { EventEmitter2 as EventEmitter } from 'eventemitter2';
 
 import symbolGenerator from 'core/symbol';
+import { readonly } from 'core/object/proxy-readonly';
 
 import { deprecate } from 'core/functools';
 import { concatURLs } from 'core/url';
@@ -370,7 +371,7 @@ export default abstract class Provider extends ParamsProvider implements IProvid
 						}
 					});
 
-					res.data = Object.freeze(composition);
+					res.data = readonly(composition);
 					return res;
 				}),